Process bigsibling metrics
We now send bigsibling metrics to fluentd on a separate port than the
normal *syslog* metrics. The format of these logs is json, rather than
syslog. Have decided to put them in the `tsuru_services` LogGroup since
these are more related to the tsuru platform itself, rather than the
applications we run.

Resolves: CTXB2S-8325